1.43 compatibility (#3)
* 1.43 compatibility

- Use new HookHandler interface
- Set up forwards-compat class aliases for older MW versions

* Add CI

* Address CI issues

- Use tabs instead of spaces
- Make unit tests use the updated hook interfaces

* Fix tests

phpunit mock methods configured with with() will disallow any parameters
not specified in the with() call, which caused tests to fail when
WebRequest::getVal was called multiple times with different parameters.
Switch to willReturnMap() to properly map certain values to return
values, while returning null for unmapped values.

* Migrate to proper unit test

This requires moving to a unit/ subdirectory to indicate that
we do not need to load MW when running these tests.

Additionally, instead of checking that we call various methods on
OutputPage when denying access, refactor denyAccess to a protected
method and mock it during testing, so that we can ensure denyAccess is
called without running any of its code. The code itself requires the
services container to be instantiated in order for wfMessage()->plain()
to work, which we cannot do during pure unit tests.

We should probably also add some integration tests to verify that
denyAccess() behaves as expected, in addition to tests that exercise
other aspects of the extension.